+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v1.5.2
+ - Fix unpausing with space also causing the player to jump
+ - Fix missing "Pan editor while swiping" editor keybind
+
 ## v1.5.1
  - Fix not being able to hold from respawn in PlayLayer